+++ sys/dev/uart/uart_16550.h
@@ -53,7 +53,7 @@ struct uart_16550_softc {
 	uint8_t reg_shift;
 };
 
-int uart_16550_init(struct uart_16550_softc *sc, uint32_t base,
+int uart_16550_init(struct uart_16550_softc *sc, size_t base,
     uint32_t uart_freq, uint32_t baud_rate, uint8_t reg_shift);
 void uart_16550_putc(struct uart_16550_softc *sc, char c);
 char uart_16550_getc(struct uart_16550_softc *sc);
